Add files via upload
developed the linear regression model, still have a lot to wrok on,but 90% accuracy is a good starting spot! had to use gpt for the comments lol, too lazy. I will continue on the next model and archtecture, that i will need to work on! i also will add in the to-do list onto the readme